Moved LogisticRegressionKernel to personal class instead of init
